-scroll tour chart with the dragged x-slider -new option in the adjust altitude dialog to select the whole tour when using srtm approach adjustment type -improved performance when drawing tour chart, best result is when tour is zoomed in

+++ net.tourbook/src/net/tourbook/mapping/TourPainter.java
@@ -241,7 +241,7 @@ public class TourPainter extends MapPainter {
 		final Color legendTextColor = display.getSystemColor(SWT.COLOR_BLACK);
 		final Color legendTextBackgroundColor = display.getSystemColor(SWT.COLOR_WHITE);
 
-		int legendValue = 0;
+		float legendValue = 0;
 
 		int unitLabelIndex = 0;
 
@@ -249,7 +249,7 @@ public class TourPainter extends MapPainter {
 
 		for (int pixelIndex = 0; pixelIndex <= availableLegendPixels; pixelIndex++) {
 
-			legendValue = (int) (legendMinValue + pixelValue * pixelIndex);
+			legendValue = legendMinValue + pixelValue * pixelIndex;
 
 			int valuePosition;
 			if (isDrawVertical) {
